backend/requirements.txt
fastapi
uvicorn[standard]

python-dotenv
pydantic

langchain
langchain-core
langchain-community
langgraph
langchain-groq
groq

matplotlib
pandas
numpy
seaborn
plotly

streamlit
python-multipart